#1.4 fix architecture test
