--- src/Core/BooksStore.Domain/Class1.cs
@@ -1,5 +1,0 @@
-﻿namespace BooksStore.Domain;
-public class Class1
-{
-
-}
