+++ src/External/Presentation/BooksStore.Debug.Console/Class1.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace BooksStore.Debug.Console;
+public class Class1
+{
+}
